latest change

+++ database/migrations/2022_10_27_051523_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('image');
             $table->string('title');
             $table->longText("description");
-            $table->float("price");
+            $table->bigInteger("price");
             $table->foreignId("category_id")->references('id')->on('categories')->onDelete('cascade');
             $table->timestamps();
             $table->softDeletes();
